refactor: move UpdateCartItemQuantityRequest to the carts package for improved organization
